requirements.txt
alembic==1.15.2
cachetools==5.5.0
fido2==1.2.0
Flask==3.1.1
Flask_Login==0.6.3
flask_mail==0.10.0
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
itsdangerous==2.2.0
python_bcrypt==0.3.2
pytz==2024.2
rapidfuzz==3.13.0
Requests==2.32.3
SQLAlchemy==2.0.36
Werkzeug==3.1.3
gunicorn==23.0.0
PyMySQL==1.1.1
